+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.0.0] - 2020-04-?
-- Rename env var AMQP_MATCHDBI -> AMQP_MATCHWRITER_UPLOAD_REPORT
+- Rename env vars
+	- AMQP_MATCHDBI -> AMQP_MATCHWRITER_UPLOAD_REPORT
+	- AMQP_SITUATIONSOPERATOR -> AMQP_SITUATIONSOPERATOR_REPORT
 
 ## [1.1.1] - 2020-04-22
 ### Added